Removing files not being used
