+++ exactraytrace/All_in_one.py
@@ -2,54 +2,6 @@ import numpy as np
 import matplotlib.pyplot as plt
 from Surfaces import *
 import Functions as Func
-
-#Useful Functions
-
-""" 
-
-The functions below will be callable libraries that can be used throughout
-The Program.  They are not designed to be included in the main program.
-
-"""
-
-""" #This arange function is designed to be more precise with floating point numbers
-def safe_arange(start, stop, step, dec):
-    round_start = round(start, dec)
-    round_stop = round(stop, dec)
-    int_start = round(round_start/step)
-    int_stop = round(round_stop/step)
-
-    return step * np.arange(int_start, int_stop+1) """
-
-
-""" # These are the Surface Functions
-
-def sphere_refract_ray(y, radius, thickness,n,dz,dec):
-    
-    sag = radius - np.sqrt(radius**2 - y**2) #lens sag at y
-    z = safe_arange(sag, thickness, dz, dec)
-    
-    sin_phi1 = y/radius
-    sin_phi2 = sin_phi1/n
-    phi1 = np.arcsin(sin_phi1)
-    phi2 = np.arcsin(sin_phi2)
-    theta = phi2-phi1 
-    slope = np.tan(theta)
-    ray = (slope*(z-sag)+y)
-    
-    return ray, slope, z
-
-def plane_refract_ray(y, slope, thickness,n, z):
-    
-    theta1 = np.arctan(slope)
-    theta2 = np.arcsin(n*np.sin(theta1))
-    slope2 = np.tan(theta2)
-    ray_air = ((z-thickness)* slope2 + y)
-    
-    return ray_air """
-
-
-
 """ 
 Main Program 
 This describes the ultimate goal of this program and how it's
