+++ exactraytrace/Functions/__init__.py
@@ -1,1 +1,3 @@
 from safe_arange import *
+from Surfaces import *
+import Functions as Func
