+++ exactraytrace/Functions/safe_arange.py
@@ -1,0 +1,8 @@
+#This arange function is designed to be more precise with floating point numbers
+def safe_arange(start, stop, step, dec):
+    round_start = round(start, dec)
+    round_stop = round(stop, dec)
+    int_start = round(round_start/step)
+    int_stop = round(round_stop/step)
+
+    return step * np.arange(int_start, int_stop+1)
