--- exactraytrace/safe_arange.py
@@ -1,10 +1,0 @@
-import numpy as np
-
-def safe_arange(start, stop, step, dec):
-    round_start = round(start, dec)
-    round_stop = round(stop, dec)
-    int_start = round(round_start/step)
-    int_stop = round(round_stop/step)
-
-
-    return step * np.arange(int_start, int_stop+1)
